% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boastApp.R
\name{boastApp}
\alias{boastApp}
\title{BOAST App}
\usage{
boastApp(ui, server, config = NA)
}
\description{
Overrides the default shinyApp ui and server config
\itemize{
\item ui includes template for accessibility fixes and consistent styling
\item server config adds setup defaults
}
}
