% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\docType{data}
\name{boastPalette}
\alias{boastPalette}
\title{boastPalette}
\format{
An object of class \code{character} of length 9.
}
\usage{
boastPalette
}
\value{
A palette
}
\description{
Based on the Wong Color Blind Set
}
\examples{
showPalette(boastPalette)

}
\seealso{
\href{https://www.nature.com/articles/nmeth.1618}{https://www.nature.com/articles/nmeth.1618}

\href{https://davidmathlogic.com/colorblind/#\%23D81B60-\%231E88E5-\%23FFC107-\%23004D40}{https://davidmathlogic.com/colorblind/#\%23D81B60-\%231E88E5-\%23FFC107-\%23004D40}
}
\keyword{datasets}
