% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appUtils.R
\name{citeApp}
\alias{citeApp}
\title{citeApp}
\usage{
citeApp(metaData)
}
\arguments{
\item{metaData}{Required--use \code{getAppMeta()} to access the DESCRIPTION
file from inside the app}
}
\value{
Character string of the app's citation
}
\description{
A function which will generate how an app should be cited. This function
requires the presence of a DESCRIPTION file in the app's repository to run.
}
\examples{
citeApp(getAppMeta())

}
