% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployApp.R
\name{deployApp}
\alias{deployApp}
\title{todo:
\itemize{
\item check if rsconnect account info is set if not ask for it
\item write library manifest
\item run lintr
\item deployApp
}}
\usage{
deployApp(...)
}
\description{
todo:
\itemize{
\item check if rsconnect account info is set if not ask for it
\item write library manifest
\item run lintr
\item deployApp
}
}
