% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockerUtils.R
\name{.bindInputEvents}
\alias{.bindInputEvents}
\title{Learning Locker Statement Generation
TODO: bind events to all fixed dashboard input items}
\usage{
.bindInputEvents(session)
}
\description{
Learning Locker Statement Generation
TODO: bind events to all fixed dashboard input items
}
