% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockerUtils.R
\name{.generateStatement}
\alias{.generateStatement}
\title{Create an xAPI Statement using rLocker}
\usage{
.generateStatement(
  session,
  verb = NA,
  object = NA,
  description = NA,
  value = NA,
  extensions = NULL
)
}
\description{
Create an xAPI Statement using rLocker
}
