% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appUtils.R
\name{.setAppIdentifier}
\alias{.setAppIdentifier}
\title{Store app identifier}
\usage{
.setAppIdentifier(uuid)
}
\description{
Writes Universally Unique ID (UUID) to App DESCRIPTION file in project root.
}
\details{
Assumes property does not exist in file already.
}
