% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appUtils.R
\name{feedmanDiaconis}
\alias{feedmanDiaconis}
\title{Feedman-Diaconis Rule}
\usage{
feedmanDiaconis(x)
}
\arguments{
\item{x}{numerical data vector}
}
\value{
number
}
\description{
Returns appropriate binwidth for histograms following the Feedman-Diaconis rule.
}
