% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockerUtils.R
\name{generateStatement}
\alias{generateStatement}
\title{Generate Statement}
\usage{
generateStatement(
  session,
  verb = NA,
  object = NA,
  description = NA,
  interactionType = NA,
  response = NULL,
  success = NA,
  score = list(),
  completion = NA,
  extensions = NULL
)
}
\description{
Create an xAPI Statement using rLocker
}
\examples{
generateStatement(
  session,
  verb = "answered",
  object = "SAMPLE_QUESTION_ID",
  description = "SAMPLE_QUESTION_OUTPUT",
  interactionType = "choice",
  response = "SAMPLE_RESPONSE",
  success = c(TRUE, FALSE),
  score = list(
    min = 0,
    max = 100,
    raw = 35,
    scaled = 0.35
  ),
  completion = FALSE
)

}
\seealso{
\link[rlocker]{getVerbList()}

\link[rlocker]{getInteractionTypes()}
}
