% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appUtils.R
\name{getAppRoot}
\alias{getAppRoot}
\title{Returns the working directory set at startup}
\usage{
getAppRoot()
}
\description{
Returns the working directory set at startup
}
\seealso{
Sys.getenv("PWD") on shinyapps.io
}
