% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appUtils.R
\name{getCurrentAddress}
\alias{getCurrentAddress}
\title{getCurrentAddress}
\usage{
getCurrentAddress(session)
}
\value{
url
}
\description{
Returns current page address in session$clientData as validated url.
Example: "https://psu-eberly.shinyapps.io/Sample_App/#shiny-tab-challenge"
}
