% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockerUtils.R
\name{getLockerConfig}
\alias{getLockerConfig}
\title{getLockerConfig}
\usage{
getLockerConfig()
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  base_url,
  auth,
  agent = list(
    name,
    mbox,
    objectType
  ),
  language
)
}\if{html}{\out{</div>}}
}
\description{
Return Learning Locker configuration for current session.
}
\seealso{
\code{\link[rLocker]{get_locker_config()}}
}
