% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{initAppStorage}
\alias{initAppStorage}
\title{Determine which table/document to use based on environment}
\usage{
initAppStorage(conn)
}
\description{
Determine which table/document to use based on environment
}
