% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appUtils.R
\name{isLocal}
\alias{isLocal}
\title{isLocal}
\usage{
isLocal()
}
\value{
boolean
}
\description{
Check to see if current environment is local or shinyapps.io.
Returns \code{TRUE} if local; \code{FALSE} if shinyapps.io.
}
