% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lintr.R
\name{linters}
\alias{linters}
\alias{boast_comment_linter}
\title{linters}
\usage{
boast_comment_linter(source_file)
}
\description{
Contains a set linters for code deployment and review.
}
\section{Functions}{
\itemize{
\item \code{boast_comment_linter}: checks that '<-' is always used for assignment
}}

