% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{logo_colors}
\alias{logo_colors}
\title{PennState Logo Colors}
\usage{
logo_colors(style = "default")
}
\arguments{
\item{style}{Logo style \code{[default, reversed, grayscale]}}
}
\value{
list
}
\description{
PennState Logo Colors
}
\seealso{
\href{http://visualidentity.psu.edu/}{http://visualidentity.psu.edu/}
}
