% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{renderIcon}
\alias{renderIcon}
\title{Icons for Apps}
\usage{
renderIcon(icon = "default", width = 36, html = FALSE)
}
\arguments{
\item{icon}{Icon style \code{[default, correct, incorrect, partial]}}

\item{width}{Icon width (in pixels)}

\item{html}{FALSE to render as shinyUI or TRUE to return html markup \code{[FALSE, TRUE]}}
}
\value{
<shiny.render.function> or html
}
\description{
Icons for Apps
}
\examples{
renderIcon("correct")
renderIcon("incorrect", width = 25, html = TRUE)

}
