% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{showPalette}
\alias{showPalette}
\title{showPalette}
\usage{
showPalette(palette = boastPalette)
}
\value{
A pie chart of palette colors.
}
\description{
Used to quickly visualize what color options are available in a given palette.
}
