% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{surveyLink}
\alias{surveyLink}
\title{surveyLink}
\usage{
surveyLink(name = NA, markup = TRUE)
}
\arguments{
\item{name}{Short Name for survey reference}

\item{markup}{Return link as markup or plain-text}
}
\value{
UI Element or Text
}
\description{
surveyLink
}
