% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boast-app.R
\name{boastApp}
\alias{boastApp}
\title{BOAST App}
\usage{
boastApp(ui, server)
}
\description{
Overrides the default shinyApp ui and server config
- ui includes template for accessibility fixes and consistent styling
- server config adds setup defaults
}
