% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\docType{data}
\name{boastPalette}
\alias{boastPalette}
\title{boastPalette}
\format{An object of class \code{character} of length 9.}
\usage{
boastPalette
}
\description{
Based on the Wong color blind set
}
\examples{
showPalette(boastPalette)

}
\keyword{datasets}
