% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployApp.R
\name{deployApp}
\alias{deployApp}
\title{todo:
- check if rsconnect account info is set if not ask for it
- write library manifest
- run lintr
- deployApp}
\usage{
deployApp(...)
}
\description{
todo:
- check if rsconnect account info is set if not ask for it
- write library manifest
- run lintr
- deployApp
}
