% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockerUtils.R
\name{generateStatement}
\alias{generateStatement}
\title{Generate Statement}
\usage{
generateStatement(
  session,
  verb = NA,
  object = NA,
  description = NA,
  interactionType = NA,
  response = NULL,
  success = NA,
  completion = NA,
  extensions = NULL
)
}
\description{
Create an xAPI Statement using rLocker
}
