% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appUtils.R
\name{getAppRoot}
\alias{getAppRoot}
\title{getAppRoot}
\usage{
getAppRoot()
}
\value{
path
}
\description{
Returns the working directory set at startup.
}
\details{
Default deployed path is \verb{/srv/connect/apps/<DIRECTORY_NAME>}.
Example: "/srv/connect/apps/Sample_App"
}
\seealso{
Sys.getenv("PWD") on shinyapps.io
}
