% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockerUtils.R
\name{getLockerConfig}
\alias{getLockerConfig}
\title{getLockerConfig}
\usage{
getLockerConfig()
}
\value{
\preformatted{list(
  base_url,
  auth,
  agent = list(
    name,
    mbox,
    objectType
  ),
  language
)
}
}
\description{
Return Learning Locker configuration for current session.
}
\seealso{
\link[rLocker]{get_locker_config()}
}
