% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\docType{data}
\name{psuPalette}
\alias{psuPalette}
\title{psuPalette}
\format{An object of class \code{character} of length 8.}
\usage{
psuPalette
}
\description{
Based on Penn State's three official color palettes.
}
\examples{
showPalette(psuPalette)

}
\seealso{
\link{https://brand.psu.edu/design-essentials.html#color}
}
\keyword{datasets}
