% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{renderIcon}
\alias{renderIcon}
\title{Icons for Apps}
\usage{
renderIcon(icon = "default", width = 50)
}
\arguments{
\item{style}{Icon style \code{[default, valid, partial, invalid]}}
}
\value{
shinyUI
}
\description{
Icons for Apps
}
\examples{
renderIcon("valid")

}
